backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
python-dotenv==1.0.0
reportlab==4.0.0
google-generativeai==0.16.0
